backend/requirements.txt
# Core server
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# HTTP client (used by AI Studio routes for Pollinations.ai)
httpx>=0.24.0

# AI API client (Replicate)
replicate>=0.25.0

# Local vision pipeline (existing features)
ultralytics>=8.0.0
simple-lama-inpainting>=0.1.0
opencv-python>=4.8.0
Pillow>=9.0.0
numpy>=1.24.0
